Atualização Final

+++ app/Http/Controllers/AddressController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Address;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class AddressController extends Controller
 {
@@ -16,8 +17,7 @@ class AddressController extends Controller
     }
 
     //Método para criar um novo endereço
-    public function create(Request $request)
-    {
+    public function create(Request $request) {
         $address = new Address();
 
         $address->zip = $request->zip;
@@ -39,8 +39,19 @@ class AddressController extends Controller
     }
 
     //Método para deletar um endereço
-    public function delete($id)
-    {
+    public function delete($id) {
+        //Verificar se existe um pedido em aberto com esse endereço. Se sim, impedir o processo de deletar
+        $isInUse = DB::table('orders')
+        ->join('order_statuses', 'orders.order_status_id', '=', 'order_statuses.id')
+        ->where([['address_id', '=', $id], ['order_statuses.final', 0]])
+        ->exists();
+
+        if ($isInUse) {
+            return response()->json([
+                "message" => "Endereço está em uso em um pedido em aberto e não pode ser deletado até sua finalização."
+            ], 403);
+        }
+
         if (Address::where('id', $id)->exists()) {
             $address = Address::find($id);
             $address->delete();
@@ -48,6 +59,32 @@ class AddressController extends Controller
             return response()->json([
                 "message" => "Endereço deletados"
             ], 202);
+        } else {
+            return response()->json([
+                "message" => "Endereço não encontrado"
+            ], 404);
+        }
+    }
+
+    //Atualizar informações do Endereço
+    public function update(Request $request, $id){
+        if (Address::where('id', $id)->exists()) {
+            $address = Address::find($id);
+
+            $address->zip = is_null($request->zip) ? $address->zip : $request->zip;
+            $address->street = is_null($request->street) ? $address->street : $request->street;
+            $address->district = is_null($request->district) ? $address->district : $request->district;
+            $address->city = is_null($request->city) ? $address->city : $request->city;
+            $address->number = is_null($request->number) ? $address->number : $request->number;
+            $address->complement = is_null($request->complement) ? $address->complement : $request->complement;
+            $address->reference = is_null($request->reference) ? $address->reference : $request->reference;
+            $address->identification = is_null($request->identification) ? $address->identification : $request->identification;
+
+            $address->save();
+
+            return response()->json([
+                "message" => "Atualizado com sucesso"
+            ], 200);
         } else {
             return response()->json([
                 "message" => "Endereço não encontrado"
